drugi

+++ v7-backend/src/main/java/rs/ac/uns/ftn/svtvezbe07/model/dto/ImageDTO.java
@@ -18,14 +18,14 @@ import rs.ac.uns.ftn.svtvezbe07.model.entity.User;
 @Getter
 @Setter
 public class ImageDTO {
-	private Integer id;	
+	private Long id;	
 	private String path;
 	private Post post;
 	private User user;
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getPath() {
@@ -52,7 +52,7 @@ public class ImageDTO {
 	public ImageDTO(Image image) {
 		this(image.getId(),image.getPath(),image.getPost(),image.getUser());
 	}
-	public ImageDTO(Integer id, String path, Post post, User user) {
+	public ImageDTO(Long id, String path, Post post, User user) {
 		super();
 		this.id = id;
 		this.path = path;
